+++ src/js/validator/bic.js
@@ -2,19 +2,12 @@
     $.fn.bootstrapValidator.i18n = $.extend(true, $.fn.bootstrapValidator.i18n || {}, {
         'en_US': {
             bic: {
-                'default': 'Please enter a valid BIC code',
-                invalidChars : 'The first 6 characters must be letters',
-                leadingZero: 'The number "0" was entered, but this should most likely be the letter "O"'
+                'default': 'Please enter a valid BIC number'
             }
         }
     });
 
     $.fn.bootstrapValidator.validators.bic = {
-        html5Attributes: {
-            message: 'message'
-        },
-
-
         /**
          * Validate an Business Identifier Code (BIC), also known as ISO 9362, SWIFT-BIC, SWIFT ID or SWIFT code
          *
@@ -33,36 +26,7 @@
             if (value === '') {
                 return true;
             }
-
-            // For user convenience lower case letter are allowed
-            value = value.toUpperCase();
-
-            var locale = validator.getLocale();
-            var message = $.fn.bootstrapValidator.i18n[locale].bic.default;
-            var valid = false;
-
-            if (/^[A-Z]{6}[A-Z0-9]{2}([A-Z0-9]{3})?$/.test(value)) {
-                valid = true;
-            } else if (!/^[A-Z]{6}.*$/.test(value)){
-                // Return a more detailed validation message if one or more of the
-                // first 6 characters is invalid
-                message = $.fn.bootstrapValidator.i18n[locale].bic.invalidChars;
-
-                // 0 and O are commonly mistaken when entering BICs. If at least
-                // one of the first 6 characters is a 0 show a hint
-                if ( -1 !== value.indexOf('0') && 6 > value.indexOf('0')  ) {
-                    message = message + '. ' +  $.fn.bootstrapValidator.i18n[locale].bic.leadingZero;
-                }
-            }
-
-            if(options.message) {
-                message = options.message;
-            }
-
-            return {
-              valid: valid,
-              message: message
-            };
+            return /^[a-zA-Z]{6}[a-zA-Z0-9]{2}([a-zA-Z0-9]{3})?$/.test(value);
         }
     }
 }(jQuery));