+++ test/spec/validator/bic.js
@@ -10,8 +10,8 @@ describe('bic', function() {
 
         $('#bicForm').bootstrapValidator();
 
-        this.bv       = $('#bicForm').data('bootstrapValidator');
-        this.$bic    = this.bv.getFieldElements('bic');
+        this.bv   = $('#bicForm').data('bootstrapValidator');
+        this.$bic = this.bv.getFieldElements('bic');
     });
 
     afterEach(function() {
@@ -20,7 +20,11 @@ describe('bic', function() {
 
     it('invalid bic', function() {
         // Test some invalid BICs
-        var invalidSamples = ['ASPKAT2LXX', 'ASPKAT2LX', 'ASPKAT2LXXX1', 'DABADKK', 'RZ00AT2L303'];
+        var invalidSamples = [
+            'ASPKAT2LXX', 'ASPKAT2LX', 'ASPKAT2LXXX1', 'DABADKK', 'RZ00AT2L303',
+            // Invalid fist 6 characters
+            '1SBACNBXSHA', 'D2BACNBXSHA', 'DS3ACNBXSHA', 'DSB4CNBXSHA', 'DSBA5NBXSHA', 'DSBAC6BXSHA', '1S3AC6BXSHA',
+        ];
         for (i in invalidSamples) {
             this.bv.resetForm();
             this.$bic.val(invalidSamples[i]);
@@ -39,52 +43,4 @@ describe('bic', function() {
             expect(this.bv.isValid()).toBeTruthy();
         }
     });
-
-    it('invalid fist 6 characters', function() {
-        // If one or more of the first 6 characters are invalid expect a special error message
-        var locale = this.bv.getLocale();
-        var invalidSamples = ['1SBACNBXSHA', 'D2BACNBXSHA', 'DS3ACNBXSHA', 'DSB4CNBXSHA', 'DSBA5NBXSHA', 'DSBAC6BXSHA', '1S3AC6BXSHA' ];
-        var expectedMessage = $.fn.bootstrapValidator.i18n[locale].bic.invalidChars
-
-        for (i in invalidSamples) {
-            this.bv.resetForm();
-            this.$bic.val(invalidSamples[i]);
-            this.bv.validate();
-
-            expect(this.bv.isValid()).toEqual(false);
-            expect(this.bv.getMessages()[0]).toEqual(expectedMessage);
-        }
-
-        // Errors after the first 6 characters should produce the standard error message
-        this.bv.resetForm();
-        this.$bic.val('ASPKAT2LXXXX');
-        this.bv.validate();
-        expect(this.bv.isValid()).toEqual(false);
-        expect(this.bv.getMessages()[0]).toEqual($.fn.bootstrapValidator.i18n[locale].bic.default);
-
-    });
-
-    it('first 6 characters contains 0', function() {
-        // If one or more of the first 6 characters are 0 expect a special error message
-        var locale = this.bv.getLocale();
-        var invalidSamples = ['0ZOOAT2L303', 'R0OOAT2L303', 'RZ00AT2L303', 'RZOO002L303' ];
-        var expectedMessage = $.fn.bootstrapValidator.i18n[locale].bic.invalidChars + '. ' + $.fn.bootstrapValidator.i18n[locale].bic.leadingZero;
-
-        for (i in invalidSamples) {
-            this.bv.resetForm();
-            this.$bic.val(invalidSamples[i]);
-            this.bv.validate();
-
-            expect(this.bv.isValid()).toEqual(false);
-            expect(this.bv.getMessages()[0]).toEqual(expectedMessage);
-        }
-
-        // Errors after the first 6 characters should produce the standard error message
-        this.bv.resetForm();
-        this.$bic.val('ASPKAT2LXXXX');
-        this.bv.validate();
-        expect(this.bv.isValid()).toEqual(false);
-        expect(this.bv.getMessages()[0]).toEqual($.fn.bootstrapValidator.i18n[locale].bic.default);
-    });
-
 });