+++ CHANGELOG.md
@@ -7,6 +7,7 @@ __New Features__
 * [#1131](https://github.com/nghuuphuoc/bootstrapvalidator/issues/1131): Support add-on
 * [#1140](https://github.com/nghuuphuoc/bootstrapvalidator/pull/1140): Add UAE [phone number](http://bootstrapvalidator.com/validators/phone/) validator, thanks to [@s-a-y](https://github.com/s-a-y)
 * [#1153](https://github.com/nghuuphuoc/bootstrapvalidator/pull/1153): Add EIN validator, thanks to [@joshuachestang](https://github.com/joshuachestang)
+* [#1165](https://github.com/nghuuphuoc/bootstrapvalidator/pull/1165): Add BIC (ISO 9362) validator, thanks to [@thomaslhotta](https://github.com/thomaslhotta)
 
 __Add-ons__
 * [#1116](https://github.com/nghuuphuoc/bootstrapvalidator/issues/1116): Showing only one message each time